+++ src/ModelSearchAspect.php
@@ -2,6 +2,7 @@
 
 namespace Spatie\Searchable;
 
+use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Illuminate\Foundation\Auth\User;
 use Illuminate\Database\Eloquent\Model;
@@ -42,12 +43,22 @@ class ModelSearchAspect extends SearchAspect
 
         if (is_array($attributes)) {
             $this->attributes = SearchableAttribute::createMany($attributes);
+
+            return;
+        }
+
+        if (is_string($attributes)) {
+            $this->attributes = SearchableAttribute::create($attributes);
+
+            return;
         }
 
         if (is_callable($attributes)) {
             $callable = $attributes;
 
             $callable($this);
+
+            return;
         }
     }
 
@@ -91,14 +102,20 @@ class ModelSearchAspect extends SearchAspect
 
     protected function addSearchConditions(Builder $query, string $term)
     {
-        foreach ($this->attributes as $attribute) {
-            $sql = "LOWER({$attribute->getAttribute()}) LIKE ?";
+        $attributes = $this->attributes;
+        $searchTerms = explode(' ', $term);
 
-            $term = mb_strtolower($term, 'UTF8');
+        $query->where(function (Builder $query) use ($attributes, $term, $searchTerms) {
+            foreach (Arr::wrap($attributes) as $attribute) {
+                foreach ($searchTerms as $searchTerm) {
+                    $sql = "LOWER({$attribute->getAttribute()}) LIKE ?";
+                    $searchTerm = mb_strtolower($searchTerm, 'UTF8');
 
-            $attribute->isPartial()
-                ? $query->orWhereRaw($sql, ["%{$term}%"])
-                : $query->orWhere($attribute->getAttribute(), $term);
-        }
+                    $attribute->isPartial()
+                        ? $query->orWhereRaw($sql, ["%{$searchTerm}%"])
+                        : $query->orWhere($attribute->getAttribute(), $searchTerm);
+                }
+            }
+        });
     }
 }